+++ src/main/java/duke/handler/FileHandler.java
@@ -1,3 +1,8 @@
+package duke.handler;
+
+import duke.task.*;
+import duke.time.Time;
+
 import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
