+++ src/main/java/duke/time/Time.java
@@ -1,3 +1,5 @@
+package duke.time;
+
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
